requirements.txt
python-dotenv
pyautogui
pyperclip
pydirectinput
pygetwindow
opencv-python
allure-pytest
pytest